Create BasicFlashingSign.py
Flashes the text red, then moves the cursor back, replaces the text with enough blank spaces (" ") to cover the text, then shows the text as green.

This program replaces the default end="\n" inside the print( ) statement with end ="\r" to make it work.

The sleep( ) is how long you show the text for before it disappears and then reappears as a different color.