Create personal_info.py
I THINK this is closure? Anyway, you cannot access the variable in the outer scope (ComanyStatements) unless (in this case), you provide the right id.

The validation is with the conditional: if not CompanyStatements.get(_id). If False, it raises an error. Otherwise, it will return the value of CompanyStatements.