Create PasswordApp.py
This app simulates using the .env library to HIDE sensitive data.

PLEASE READ THE COMMENT ABOVE THIS FILE OR IT MAY NOT WORK WHEN YOU CLONE/RUN THIS FILE!!!

This is probably what you would do if, say, you either buy or use an API that involves the owner of that API giving you an API key.

Traditionally, you would not want to store that API key in the variable of your (in this case) Password.App file (yes... I did put the password inside the function argument, but only as a demo!!!).

Instead, you'd want to... JUST READ THE COMMENT!!!