filemanager.
Automatically file.closes( ) a file, so you do not have to always remember how to do file.close( ).

YES... I know Python already has this, but still, this is kind of like a recreation of how the native with open( filename ) as f works.

It basically yields the open( file, mode ) as an ALIAS (shown here as "f") so you can do (almost) whatever you want.