Changed the __add__ method.
The __add__ method now adds a 0.1 tax to Order and returns an Order object:

Order(1, (item_1_total + item_2_total), True)

This is so I can add multiple Orders together: Order( ... ) + Order ( ... ) + Order ( ... ).

If I were to return, say, an integer (which I could), then, if I were to do Order( ... ) + Order ( ....) + Order(...), it would throw an error when it tries to add the third Order(...).

WHY?

If I returned a number instead of Order(...), then:
Order(...) + Order(...) would return an integer or float. So far, no errors...

BUT...

That integer/float would be added to the third Order(...).

You cannot add type(int) to type(Object), which is what Order( ... ) is.